avoid mac make error

+++ examples/bootstrap_server/bootstrap_server.c
@@ -88,7 +88,7 @@ void print_usage(const char *filename, const char *port) {
     fprintf(stdout, "  -l PORT\tSet the local UDP port of the Bootstrap Server. Default: %s\r\n", port);
     fprintf(stdout, "  -4\t\tUse IPv4 connection. Default: IPv6 connection\r\n");
     fprintf(stdout, "  -S BYTES\tCoAP block size. Options: 16, 32, 64, 128, 256, 512, 1024. Default: %" PRIu16 "\r\n",
-            LWM2M_COAP_DEFAULT_BLOCK_SIZE);
+            (uint16_t)LWM2M_COAP_DEFAULT_BLOCK_SIZE);
     fprintf(stdout, "\r\n");
 }
